backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
supabase==2.0.2
python-dotenv==1.0.0
pydantic==2.5.0
python-multipart==0.0.6
httpx>=0.24.0,<0.25.0
# AWS S3 dependencies removed - now using Supabase Storage
pdfplumber==0.11.4
PyPDF2==3.0.1
python-docx==1.1.0
openai>=1.6.1
Pillow==10.3.0
pytesseract==0.3.10
razorpay==1.3.0
PyJWT==2.8.0
# Email dependencies
aiosmtplib==3.0.1
jinja2==3.1.2
# Additional dependencies for teacher verification
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4